1,clean the warning 2,add the code license 3,move the pppos port to the component
