+++ components/net/ppp_port/osport/ring.c
@@ -1,3 +1,36 @@
+/*----------------------------------------------------------------------------
+ * Copyright (c) <2016-2018>, <Huawei Technologies Co., Ltd>
+ * All rights reserved.
+ * Redistribution and use in source and binary forms, with or without modification,
+ * are permitted provided that the following conditions are met:
+ * 1. Redistributions of source code must retain the above copyright notice, this list of
+ * conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright notice, this list
+ * of conditions and the following disclaimer in the documentation and/or other materials
+ * provided with the distribution.
+ * 3. Neither the name of the copyright holder nor the names of its contributors may be used
+ * to endorse or promote products derived from this software without specific prior written
+ * permission.
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
+ * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+ * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
+ * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+ * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+ * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
+ * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
+ * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *---------------------------------------------------------------------------*/
+/*----------------------------------------------------------------------------
+ * Notice of Export Control Law
+ * ===============================================
+ * Huawei LiteOS may be subject to applicable export control laws and regulations, which might
+ * include those applicable to Huawei LiteOS of U.S. and the country in which you are located.
+ * Import, export and usage of Huawei LiteOS in any manner by you shall be in compliance with such
+ * applicable export control laws and regulations.
+ *---------------------------------------------------------------------------*/
 #include <string.h>
 #include "osport.h"
 //ring:which to be initialized
@@ -120,7 +153,7 @@ s32_t ring_read(tagRingBuf *ring,u8_t *buf, s32_t len)
 	}
 	return ret;
 }
-
+//check how many data in the ring while -1 means someting error
 s32_t ring_datalen(tagRingBuf *ring)
 {
 	int ret = -1;
@@ -130,6 +163,7 @@ s32_t ring_datalen(tagRingBuf *ring)
 	}
 	return ret;
 }
+//reset the ring
 s32_t ring_reset(tagRingBuf *ring)
 {
 	int ret = -1;
@@ -141,5 +175,15 @@ s32_t ring_reset(tagRingBuf *ring)
 	}
 	return ret;
 }
-
+//deinit the ring:-1 means something error while 0 ok
+s32_t ring_deinit(tagRingBuf *ring)
+{
+	int ret = -1;
+	if(NULL != ring)
+	{
+        memset(ring,0,sizeof(tagRingBuf));
+        ret = 0;
+	}
+	return ret;
+}
